Support naming transforms on the command line.
This uses the tag name as the transformation name for the CLI.  The
values are now stored in a Dictionary (name as key, operator as value)
for random access.

Command line arguments are split by whether they are a valid
transformation name or not; then, all listed transformations are applied
to all listed files.

Obvious regression: no files OR no transforms will currently crash the
program...