Channel type rewrite, huge disaster edition.
Channels now know their range, because the I and Q channels of YIQ are
balanced around 0, not 0.0 to 1.0.  However, _something_ is broken
because the results are a mess (HSV comes out black; YIQ is...
backwards somehow) and I don't know yet what it is.  I'm just going to
leave it for today and check it again later/tomorrow.