`nearbits` allowing I/Q channels to become 0.
Since I/Q are on the interval `[-x,+x]`, an even number of quantization
levels will not allow 0 as a post-quantization output.  It's either plus
or minus a half-quantum.

This commit introduces `nearbits` which is not `2**depth` levels, but
`2**depth - 1`.  This increases the quantum slightly, but allows for 0
in the output.

The shortcuts like `nearbits 5` are `near5`.

This is still producing wild color fringing, though.  Not sure what's up
with that, yet.