Clamp Channel values to range on construction.
Previously, quantizing in the alternate color space could result in RGB
values below 0.0 or above 1.0.  When these get multiplied by 255 and
converted to bytes, the system's `byte` conversion seems to use the
value modulo 256 instead of throwing an exception.

The `Channel.Std` constructor now saturates the value instead.  This
fixes the color fringing in YUV and YIQ.