Rewrite! Every colorspace gets a pixel type!
We redefine Pixel as PxRGB, and add PxYUV, PxYIQ, PxHSV, PxGray.

We add two new interfaces: Channel for the main channel operations (raw,
normalize, etc.) and MatrixPixel for pixel types that can be sensibly
transformed linearly.

That linear transform has replaced the Array2D-based matrix
multiplication routine with a simpler routine.  It takes a 1D float
array and a MatrixPixel, asks the latter for its float tuple form, and
hard-codes the 3x3*3x1 multiplication on those array/tuple elements.
What it loses in generality, it makes up in speed.

Channel is implemented by a few new, specific types: Channel1,
ChannelMod, and ChannelAxis.  These correspond to the former
Channel.Std, Channel.Hue, and Channel constructors.  The idea is that
each pixel type should be composed of channels that understand their own
bounds.  This lets channels that are already normalized to 1.0 skip any
sort of math in normalize(), because they can know they already are.

ChannelAxis takes a limit and maps the value to between negative limit
and positive limit (centered on 0.0.)  This is used for U/V and I/Q in
YUV and YIQ.

Many types have been made private, or given private constructors, so
that invalid state cannot be represented.

Currently, YIQ/YUV/HSV are all broken in their own special ways.
ChannelHue and ChannelAxis haven't actually been debugged; this is just
the first cut of the post-Types-rewrite code that compiles.