Micro optimizations.
Replaces match with if/else in fmod.

Makes the types private, and the constructor aliases inline (at the cost
of no longer being partial applications.)

It's worth a few small gains.  Not as much as going from Debug to
Release, though, so these are probably the end of it.

(best of 3, warm caches)