Don't block the main thread while crushing.
Run the heavy bitmap processing on a background thread.

Because the thread locks the bits, also implement a pair of flags for
running/request flags.  If we're running, we set request; after running,
we check request and re-call ourselves if one was made.

This design guarantees progress (nothing will be canceled, but
intermediate requests may be skipped) but may delay the final result
while the first processing pass completes.  Also, anyone actually trying
to use the _result_ of the processing will get a bogus answer in the
code path that sets the request flag.

I think it's safe to say, I have no idea what I'm doing.  As usual.
However, since it appears to be working at the moment, I'm committing it
now and we can hack it up later.